# RSDC-API/RSDC-API/Models/DTOs/Requests/UserLoginRequest.cs
using System.ComponentModel.DataAnnotations;

namespace RSDC_API
{
    public class UserLoginRequest
    {
        [Required]
        [EmailAddress]
        public string Email { get; set; }
        [Required]
        public string Password { get; set; }
    }
}